Refactor file paths and include statements for better organization and consistency
